refactor: use persistent table state
